requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
pyyaml>=6.0
pandas>=2.0.0
python-dotenv>=1.0.0

# HuggingFace
transformers
huggingface_hub>=0.22.0
accelerate>=0.28.0
qwen-vl-utils>=0.0.8

# APIs
mistralai>=1.0.0
openai>=1.0.0

# Optionnel — attention kernels
# flash-attn>=2.5.0
